merge_two_sorted_lists.java
Use a pointer to traverse each list, picking off the smaller of the two nodes each time and linking it to the result list.

Code; dummy is a placeholder node that points to the head of the merged list 'tail' keeps track of the last node in the result list so you can easily append to it. 

Traverse both list1 and list2, appending the smaller of the two current nodes to tail and advancing that list's pointer. When one of the lists is exhausted, attach the rest of the other list to the result. 

Then, return dummy.next as the head of the merged list, skipping over the placeholder node.