maximumAreaOfLongestDiagonalRectangle.go
You are given a list of rectangles (dimensions[i] = [length, width]). 

For each rectangle, compute the diagonal length using the formula sqrt(length² + width²).

Solution - Return the area of the rectangle that has the longest diagonal.
If multiple rectangles have the same longest diagonal, return the one with the maximum area.